chore: move custom type definitions from graphql package to util package to be available in helper functions
