+++ packages/graphql/src/services/accounts.ts
@@ -1,6 +1,6 @@
 import * as DB from '@klicker-uzh/prisma'
 import { Locale, UserLoginScope, UserRole } from '@klicker-uzh/prisma'
-import { processQuestionData } from '@klicker-uzh/util'
+import { DisplayMode, processQuestionData } from '@klicker-uzh/util'
 import bcrypt from 'bcryptjs'
 import dayjs from 'dayjs'
 import { CookieOptions } from 'express'
@@ -9,7 +9,6 @@ import { Context, ContextWithUser } from '../lib/context.js'
 import { prepareInitialQuestionInstanceResults } from '../lib/questions.js'
 import { sendTeamsNotifications } from '../lib/util.js'
 import * as EmailService from '../services/email.js'
-import { AllQuestionTypeData, DisplayMode } from '../types/app.js'
 
 const COOKIE_SETTINGS: CookieOptions = {
   domain: process.env.COOKIE_DOMAIN,
@@ -1276,9 +1275,9 @@ async function seedDemoQuestions(ctx: ContextWithUser) {
                 order: ix,
                 type: DB.QuestionInstanceType.SESSION,
                 pointsMultiplier: 2 * question.pointsMultiplier,
-                questionData: processedQuestionData as AllQuestionTypeData,
+                questionData: processedQuestionData!,
                 results: prepareInitialQuestionInstanceResults(
-                  processedQuestionData as AllQuestionTypeData
+                  processedQuestionData!
                 ),
                 question: {
                   connect: { id: question.id },
