+++ packages/graphql/src/services/microLearning.ts
@@ -8,10 +8,10 @@ import {
   getInitialElementResults,
   getInitialInstanceStatistics,
   processElementData,
+  StackInput,
 } from '@klicker-uzh/util'
 import dayjs from 'dayjs'
 import { GraphQLError } from 'graphql'
-import { AllElementTypeData, StackInput } from 'src/types/app.js'
 import { v4 as uuidv4 } from 'uuid'
 import { Context, ContextWithUser } from '../lib/context.js'
 import { computeStackEvaluation } from './practiceQuizzes.js'
@@ -264,9 +264,7 @@ export async function manipulateMicroLearning(
           elements: {
             create: stack.elements.map((elem) => {
               const element = elementMap[elem.elementId]!
-              const processedElementData = processElementData(
-                element
-              ) as AllElementTypeData
+              const processedElementData = processElementData(element)
               const initialResults = getInitialElementResults(element)
 
               return {
