+++ packages/util/tsconfig.json
@@ -15,6 +15,7 @@
     "strict": true,
     "noUncheckedIndexedAccess": true,
     "noImplicitOverride": true,
+    "noErrorTruncation": true,
     /* If transpiling with TypeScript: */
     "module": "NodeNext",
     "outDir": "dist",
